% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAPSIM.R
\docType{data}
\name{all_apsim}
\alias{all_apsim}
\alias{GE_apsim}
\alias{nGE_apsim}
\alias{GnE_apsim}
\alias{nGnE_apsim}
\title{all_apsim}
\format{
A data frame with 33278 rows and 63 variables:
\describe{
\item{\code{groups}}{integer COLUMN_DESCRIPTION}
\item{\code{Env}}{integer COLUMN_DESCRIPTION}
\item{\code{trial}}{integer COLUMN_DESCRIPTION}
\item{\code{sowing}}{double COLUMN_DESCRIPTION}
\item{\code{geno}}{integer COLUMN_DESCRIPTION}
\item{\code{yld}}{double COLUMN_DESCRIPTION}
\item{\code{p1_biomassmean}}{double COLUMN_DESCRIPTION}
\item{\code{p1_epmean}}{double COLUMN_DESCRIPTION}
\item{\code{p1_photomean}}{double COLUMN_DESCRIPTION}
\item{\code{p1_radmean}}{double COLUMN_DESCRIPTION}
\item{\code{p1_radsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p1_rainsum}}{double COLUMN_DESCRIPTION}
\item{\code{p1_tstress}}{integer COLUMN_DESCRIPTION}
\item{\code{p1_ttsum}}{double COLUMN_DESCRIPTION}
\item{\code{p1_vtsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p1_wbnsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p1_wbpsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p2_biomassmean}}{double COLUMN_DESCRIPTION}
\item{\code{p2_epmean}}{double COLUMN_DESCRIPTION}
\item{\code{p2_photomean}}{double COLUMN_DESCRIPTION}
\item{\code{p2_radmean}}{double COLUMN_DESCRIPTION}
\item{\code{p2_radsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p2_rainsum}}{double COLUMN_DESCRIPTION}
\item{\code{p2_tstress}}{integer COLUMN_DESCRIPTION}
\item{\code{p2_ttsum}}{double COLUMN_DESCRIPTION}
\item{\code{p2_wbnsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p2_wbpsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p3_biomassmean}}{double COLUMN_DESCRIPTION}
\item{\code{p3_epmean}}{double COLUMN_DESCRIPTION}
\item{\code{p3_photomean}}{double COLUMN_DESCRIPTION}
\item{\code{p3_q}}{double COLUMN_DESCRIPTION}
\item{\code{p3_radmean}}{double COLUMN_DESCRIPTION}
\item{\code{p3_radsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p3_rainsum}}{double COLUMN_DESCRIPTION}
\item{\code{p3_sdmean}}{double COLUMN_DESCRIPTION}
\item{\code{p3_tstress}}{integer COLUMN_DESCRIPTION}
\item{\code{p3_ttsum}}{double COLUMN_DESCRIPTION}
\item{\code{p3_wbnsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p3_wbpsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p4_biomassmean}}{double COLUMN_DESCRIPTION}
\item{\code{p4_epmean}}{double COLUMN_DESCRIPTION}
\item{\code{p4_photomean}}{double COLUMN_DESCRIPTION}
\item{\code{p4_radmean}}{double COLUMN_DESCRIPTION}
\item{\code{p4_radsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p4_rainsum}}{double COLUMN_DESCRIPTION}
\item{\code{p4_sdmean}}{double COLUMN_DESCRIPTION}
\item{\code{p4_tstress}}{integer COLUMN_DESCRIPTION}
\item{\code{p4_ttsum}}{double COLUMN_DESCRIPTION}
\item{\code{p4_wbnsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p4_wbpsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p5_biomassmean}}{double COLUMN_DESCRIPTION}
\item{\code{p5_epmean}}{double COLUMN_DESCRIPTION}
\item{\code{p5_flowering}}{integer COLUMN_DESCRIPTION}
\item{\code{p5_photomean}}{double COLUMN_DESCRIPTION}
\item{\code{p5_radmean}}{double COLUMN_DESCRIPTION}
\item{\code{p5_radsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p5_rainsum}}{double COLUMN_DESCRIPTION}
\item{\code{p5_sdmean}}{double COLUMN_DESCRIPTION}
\item{\code{p5_tstress}}{integer COLUMN_DESCRIPTION}
\item{\code{p5_ttsum}}{double COLUMN_DESCRIPTION}
\item{\code{p5_wbnsum}}{integer COLUMN_DESCRIPTION}
\item{\code{p5_wbpsum}}{integer COLUMN_DESCRIPTION}
\item{\code{year}}{double COLUMN_DESCRIPTION}
}

An object of class \code{data.frame} with 33278 rows and 63 columns.

An object of class \code{data.frame} with 8112 rows and 63 columns.

An object of class \code{data.frame} with 6395 rows and 63 columns.

An object of class \code{data.frame} with 1560 rows and 63 columns.
}
\usage{
all_apsim

GE_apsim

nGE_apsim

GnE_apsim

nGnE_apsim
}
\description{
DESCRIPTION OF THE DATA
}
\keyword{datasets}
