% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAPSIM.R
\docType{data}
\name{apsim_geno}
\alias{apsim_geno}
\title{apsim_geno}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 199 rows and 3035 columns.
}
\usage{
apsim_geno
}
\description{
DESCRIPTION OF THE DATA
}
\keyword{datasets}
