% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAPSIM.R
\docType{data}
\name{apsim_K}
\alias{apsim_K}
\title{apsim_K}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 199 rows and 199 columns.
}
\usage{
apsim_K
}
\description{
DESCRIPTION OF THE DATA
}
\keyword{datasets}
