% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDROPS.R
\docType{data}
\name{drops_GE}
\alias{drops_GE}
\alias{drops_GnE}
\alias{drops_nGnE}
\title{drops_GE}
\format{
A data frame with 6150 rows and 25 variables:
\describe{
\item{\code{Experiment}}{integer COLUMN_DESCRIPTION}
\item{\code{parent1}}{integer COLUMN_DESCRIPTION}
\item{\code{Code_ID}}{integer COLUMN_DESCRIPTION}
\item{\code{Variety_ID}}{integer COLUMN_DESCRIPTION}
\item{\code{Accession_ID}}{integer COLUMN_DESCRIPTION}
\item{\code{Tnight.Early}}{double COLUMN_DESCRIPTION}
\item{\code{Tnight.Flo}}{double COLUMN_DESCRIPTION}
\item{\code{Tnight.Fill}}{double COLUMN_DESCRIPTION}
\item{\code{Ri.Early}}{double COLUMN_DESCRIPTION}
\item{\code{Ri.Flo}}{double COLUMN_DESCRIPTION}
\item{\code{Ri.Fill}}{double COLUMN_DESCRIPTION}
\item{\code{Psi.Flo}}{double COLUMN_DESCRIPTION}
\item{\code{Psi.Fill}}{double COLUMN_DESCRIPTION}
\item{\code{Tmax.Early}}{double COLUMN_DESCRIPTION}
\item{\code{Tmax.Flo}}{double COLUMN_DESCRIPTION}
\item{\code{Tmax.Fill}}{double COLUMN_DESCRIPTION}
\item{\code{grain.yield}}{double COLUMN_DESCRIPTION}
\item{\code{grain.number}}{double COLUMN_DESCRIPTION}
\item{\code{seed.size}}{double COLUMN_DESCRIPTION}
\item{\code{anthesis}}{double COLUMN_DESCRIPTION}
\item{\code{silking}}{double COLUMN_DESCRIPTION}
\item{\code{plant.height}}{double COLUMN_DESCRIPTION}
\item{\code{tassel.height}}{double COLUMN_DESCRIPTION}
\item{\code{ear.height}}{double COLUMN_DESCRIPTION}
\item{\code{type}}{character COLUMN_DESCRIPTION}
}

An object of class \code{data.frame} with 384 rows and 25 columns.

An object of class \code{data.frame} with 224 rows and 22 columns.
}
\usage{
drops_GE

drops_GnE

drops_nGnE
}
\description{
DESCRIPTION OF THE DATA
}
\keyword{datasets}
