% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDROPS.R
\docType{data}
\name{drops_K}
\alias{drops_K}
\title{drops_K}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 302 rows and 302 columns.
}
\usage{
drops_K
}
\description{
DESCRIPTION OF THE DATA
}
\keyword{datasets}
