% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factReg.R
\docType{package}
\name{factReg-package}
\alias{factReg}
\alias{factReg-package}
\title{factReg: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl} (\href{https://orcid.org/0000-0002-8673-2514}{ORCID})

Authors:
\itemize{
  \item Kruijer Willem (\href{https://orcid.org/0000-0001-7179-1733}{ORCID})
}

Other contributors:
\itemize{
  \item Vahe Avagyan (\href{https://orcid.org/0000-0001-5838-4850}{ORCID}) [contributor]
}

}
\keyword{internal}
