% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frBLUP.R
\name{frBLUP}
\alias{frBLUP}
\title{Penalized factorial regression on the GBLUPs}
\usage{
frBLUP(
  dat,
  Y,
  G,
  E,
  indices = NULL,
  indicesData = NULL,
  main.covariates = NULL,
  main.cov.pen = 1,
  K,
  method = c("factReg", "perGeno"),
  testEnv = NULL,
  type = NULL,
  target = c("generic", "nGnE", "GnE"),
  genPred = NULL,
  gpWeight = NULL,
  useRes = 2,
  outputFile = "results_glmnet",
  outputDir = getwd(),
  corType = c("pearson", "spearman"),
  partition = NULL,
  alpha = 1,
  lambda = NULL,
  penE = 0,
  scaling = c("no", "train", "all"),
  postLasso = FALSE,
  quadratic = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dat}{A data-frame with data from multi-environment trials.
Each row corresponds to a particular genotype in a particular environment.
The data do not need to be balanced, i.e. an environment does not need to
contain all genotypes. dat should contain the training as well as the test
environments (see testEnv)}

\item{Y}{The trait to be analyzed: either of type character, in which case
it should be one of the column names in dat, or numeric, in which case the Yth
column of dat will be analyzed.}

\item{G}{The column in dat containing the factor genotype (either character or
numeric).}

\item{E}{The column in dat containing the factor environment (either character
or numeric).}

\item{indices}{The columns in dat containing the environmental indices (vector
of type character). Alternatively, if the indices are always constant within
environments (i.e. not genotype dependent), the environmental data can also
be provided using the argument indicesData (see below).}

\item{indicesData}{An optional data-frame containing environmental indices
(covariates); one value for each environment and index. It should have the
environment names as row-names (corresponding to the names contained in dat$E);
the column names are the indices. If indices (see before) is also provided,
the latter will be ignored.}

\item{main.covariates}{Optional. The columns in dat containing main covariates,
i.e. with effects that are constant across genotypes. May overlap with
indices.}

\item{main.cov.pen}{Numeric vector corresponding to main.covariates,
indicating the relative penalties used in glmnet. Default: all 1.}

\item{K}{kinship matrix, or a list of such matrices. (TO DO: checks)}

\item{method}{(TO DO) 'factReg' (default) or 'perGeno'}

\item{testEnv}{vector (character). Data from these environments are not used
for fitting the model. Accuracy is evaluated for training and test
environments separately. The default is NULL, i.e. no test environments,
in which case the whole dataset is training. It is also possible that
there are test environments, but without any data; in this case, no accuracy
is reported for test environments (CHECK correctness).}

\item{type}{Character vector of length nrow(dat), which whould indicate the
type of observation, which should be one of 'GE', 'GnE', 'nGE', 'nGnE'}

\item{target}{character; should be either 'GnE' or 'nGnE' (default)}

\item{genPred}{vector of length nrow(dat); default NULL}

\item{gpWeight}{vector of length nrow(dat); default NULL}

\item{useRes}{If method is 'perGeno': Indicates whether the genotype-specific regressions are to be
fitted on the residuals of a model with main effects. 0 means no
(i.e. fit the regressions on the original data). 1:
residuals of a model with environmental main effects. 2 (default): genotypic and
environmental main effects.}

\item{outputFile}{The file name of the output files (without extensions,
which are added by the function)}

\item{outputDir}{The directory output to which output-files are to be
written.}

\item{corType}{type of correlation: Pearson (default) or spearman rank sum
(to be implemented).}

\item{partition}{data.frame with columns E and partition.}

\item{alpha}{type of penalty, as in glmnet (1 = LASSO, 0 = ridge; in between
= elastic net).}

\item{lambda}{Numeric vector; defines the grid over which the penalty lambda
is optimised in cross validation.
Default: NULL (defined by glmnet).
Important special case: lambda = 0 (no penalty).
May be better handled with lm rather than glmnet (ask Vahe)}

\item{penE}{numeric; default 0. If 1, environmental main effects are
penalized. If 0, they are not. Any nonnegative real number is allowed}

\item{scaling}{determines how the environmental variables are scaled.
"train" : all data (test and training environments) are scaled
using the mean and and standard deviation in the training environments.
"all" : using the mean and standard deviation of all environments.
'no' : No scaling}

\item{postLasso}{boolean.}

\item{quadratic}{boolean. If TRUE, quadratic terms (i.e., squared indices) are
added to the model. Only for those indices that were selected (CLARIFY)}

\item{verbose}{boolean. If TRUE, the accuracies per environment are printed on screen}
}
\value{
A list with the following elements:
\describe{
\item{predTrain}{Vector with predictions for the training set (to do: Add the
factors genotype and environment; make a dataframe)}
\item{predTest}{Vector with predictions for the test set (to do: Add the factors
genotype and environment; make a dataframe). To do: add estimated environmental main effects, not only predicted environmental main effects}
}
}
\description{
Based on multi-environment field trials, fits the factorial regression model
\deqn{G_{ij} = \mu + e_j + g_i + \sum_{k=1}^s \beta_{ik} x_{ij} + \epsilon_{ij},}
where .... with environmental main effects $e_j$, genotypic main effects \eqn{g_{i}} and
genotype-specific environmental sensitivities \eqn{\beta_{ik}}. See e.g. Millet
et al 2019 and van Eeuwijk / Denis 1996 (?). There are \eqn{s} environmental
indices with values \eqn{x_{ij}}. Optionally, predictions can be made for a set
of test environments, for which environmental indices are available. The new
environments must contain the same set of genotypes, or a subset
(For genomic prediction of new genotypes, the function ...).
Constraints: the first level of the factor genotype is put to 0.
Penalization: model (...) is fitted using glmnet, simultaneously penalizing
\eqn{e_j + g_i} and \eqn{\beta_{ik}}. If penG = FALSE (or penE = FALSE) the
main effects \eqn{g_i} (respectively \eqn{e_j}) are not been penalised.
Predictions for the test environments are first constructed using the
estimated main effects and sensitivities; next, predicted environmental
main effects are added. The latter are obtained by regressing the estimated
environmental main effects for the training environment on the average values
of the indices in these environments, as in Millet et al 2019.
}
