% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GnE.R
\name{GnE}
\alias{GnE}
\title{Penalized factorial regression using glmnet}
\usage{
GnE(
  dat,
  Y,
  G,
  E,
  indices = NULL,
  indicesData = NULL,
  mainCovariates = NULL,
  mainCovPen = 1,
  testEnv = NULL,
  weight = NULL,
  outputFile = "results_glmnet",
  outputDir = getwd(),
  corType = c("pearson", "spearman"),
  partition = data.frame(),
  nfolds = NULL,
  alpha = 1,
  lambda = NULL,
  penG = 0,
  penE = 0,
  scaling = c("train", "all", "no"),
  postLasso = FALSE,
  quadratic = FALSE,
  genoAcc = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dat}{A \code{data.frame} with data from multi-environment trials.
Each row corresponds to a particular genotype in a particular environment.
The data do not need to be balanced, i.e. an environment does not need to
contain all genotypes. \code{dat} should contain the training as well as the
test environments (see testEnv)}

\item{Y}{The trait to be analyzed: either of type character, in which case
it should be one of the column names in \code{dat}, or numeric, in which
case the Yth column of \code{dat} will be analyzed.}

\item{G}{The column in \code{dat} containing the factor genotype (either
character or numeric).}

\item{E}{The column in \code{dat} containing the factor environment
(either character or numeric).}

\item{indices}{The columns in \code{dat} containing the environmental
indices (vector of type character). Alternatively, if the indices are always
constant within environments (i.e. not genotype dependent), the
environmental data can also be provided using the argument \code{indicesData}
(see below).}

\item{indicesData}{An optional \code{data.frame} containing environmental
indices (covariates); one value for each environment and index. It should
have the environment names as row names (corresponding to the names
contained in \code{dat$E}); the column names are the indices. If
\code{indices} (see before) is also provided, the latter will be ignored.}

\item{mainCovariates}{Optional. The columns in \code{dat} containing main
covariates, i.e. with effects that are constant across genotypes. May
overlap with indices.}

\item{mainCovPen}{Numeric vector corresponding to \code{mainCovariates},
indicating the relative penalties used in glmnet. Default: all 1.}

\item{testEnv}{vector (character). Data from these environments are not used
for fitting the model. Accuracy is evaluated for training and test
environments separately. The default is \code{NULL}, i.e. no test
environments, in which case the whole dataset is training. It is also
possible that there are test environments, but without any data; in this
case, no accuracy is reported for test environments (CHECK correctness).}

\item{weight}{Numeric vector of length \code{nrow(dat)}, specifying the
weight (inverse variance) of each observation, used in glmnet. Default
\code{NULL}, giving constant weights.}

\item{outputFile}{The file name of the output files, without .csv extension,
which is added by the function}

\item{outputDir}{The directory to which output-files are to be
written.}

\item{corType}{type of correlation: Pearson (default) or spearman rank sum.}

\item{partition}{\code{data.frame} with columns E and partition. The column
E should contain the training environments (type character); partition
should be of type integer. Environments in the same fold should have
the same integer value. Default is \code{data.frame()}, in which case the
function uses a leave-one-environment out cross-validation. If \code{NULL},
the (inner) training sets used for cross-validation will be drawn randomly
from all observations, ignoring the environment structure. In the latter
case, the number of folds (nfolds) can be specified.}

\item{nfolds}{Default \code{NULL}. If \code{partition == NULL}, this can be
used to specify the number of folds to be used in glmnet.}

\item{alpha}{Type of penalty, as in glmnet (1 = LASSO, 0 = ridge; in between
= elastic net). Default is 1.}

\item{lambda}{Numeric vector; defines the grid over which the penalty lambda
is optimised in cross validation.
Default: NULL (defined by glmnet).
Important special case: lambda = 0 (no penalty).
May be better handled with lm rather than glmnet (ask Vahe)}

\item{penG}{numeric; default 0. If 1, genotypic main effects are
penalized. If 0, they are not. Any non negative real number is allowed.}

\item{penE}{numeric; default 0. If 1, environmental main effects are
penalized. If 0, they are not. Any non negative real number is allowed.}

\item{scaling}{determines how the environmental variables are scaled.
"train" : all data (test and training environments) are scaled
using the mean and and standard deviation in the training environments.
"all" : using the mean and standard deviation of all environments.
"no" : No scaling.}

\item{postLasso}{boolean; default \code{FALSE}. Indicates whether to refit
the model without penalty, using only the variables selected by
lasso. To do: Require that alpha = 1.}

\item{quadratic}{boolean; default \code{FALSE}. If \code{TRUE}, quadratic
terms (i.e., squared indices) are added to the model. Only for those indices
that were selected (CLARIFY).}

\item{genoAcc}{character. The accuracies per environment are evaluated
using these genotypes. It must be a subset of the genotypes in \code{dat}}

\item{verbose}{boolean; default \code{FALSE}. If \code{TRUE}, the accuracies
per environment are printed on screen.}
}
\value{
A list with the following elements:
\describe{
\item{predTrain}{Vector with predictions for the training set (to do: Add
the factors genotype and environment; make a dataframe)}
\item{predTest}{Vector with predictions for the test set (to do: Add the
factors genotype and environment; make a dataframe). To do: add estimated
environmental main effects, not only predicted environmental main effects}
\item{resTrain}{Vector with residuals for the training set}
\item{resTest}{Vector with residuals for the test set}
\item{mu}{the estimated overall (grand) mean}
\item{envInfoTrain}{The estimated environmental main effects, and the
predicted effects, obtained when the former are regressed on the averaged
indices, using penalized regression.}
\item{envInfoTest}{The predicted environmental main effects for the test
environments, obtained from penalized regression using the estimated
main effects for the training environments and the averaged indices.}
\item{parGeno}{dataframe containing the estimated genotypic main effects
(first column) and sensitivities (subsequent columns)}
\item{main.par}{If the option mainCovariates is used, this vector contains
estimates of the main effects of the indices specified in mainCovariates.}
\item{trainAccuracyEnv}{a data-frame with the accuracy (r) for each
training environment, as well as the root mean squre error (RMSE), mean
abosolute deviation (MAD) and rank (the latter is a proportion: how many
of the best 5 genotypes are in the top 10). To be removed or further
developed. All these quantities are also evaluated for a model with only
genotypic and environmental main effects (columns rMain, RMSEmain and
rankMain).}
\item{testAccuracyEnv}{A data-frame with the accuracy for each test
environment, with the same columns as trainAccuracyEnv.}
\item{trainAccuracyGeno}{a data-frame with the accuracy (r) for each
genotype, averaged over the training environments}
\item{testAccuracyGeno}{a data-frame with the accuracy (r) for each
genotype, averaged over the test environments}
\item{lambda}{The value of lambda selected using cross validation.}
\item{lambdaSequence}{...}
\item{RMSEtrain}{The root mean squared error on the training environments}
\item{RMSEtest}{The root mean squared error on the test environments}
\item{Y}{The name of the trait that was predicted, i.e. the column name
in \code{dat} that was used}
\item{G}{The genotype label that was used, i.e. the argument G that was
used}
\item{E}{The environment label that was used, i.e. the argument E that
was used}
\item{indices}{The indices that were used, i.e. the argument indices that
was used}
\item{postLasso}{The postLasso option that was used}
\item{quadratic}{The quadratic option that was used}
}
}
\description{
\loadmathjax
Based on multi-environment field trials, fits the factorial regression model
\mjeqn{Y_{ij} = \mu + e_j + g_i + \sum_{k=1}^s \beta_{ik} x_{ij} + \epsilon_{ij},}{ascii}
with environmental main effects \mjeqn{e_j}{ascii}, genotypic main effects \mjeqn{g_{i}}{ascii} and
genotype-specific environmental sensitivities \mjeqn{\beta_{ik}}{ascii}. See e.g. Millet
et al 2019 and Bustos-Korts et al 2019. There are \mjeqn{s}{ascii}
environmental indices with values \mjeqn{x_{ij}}{ascii}. Optionally, predictions can be made for a set
of test environments, for which environmental indices are available. The new
environments must contain the same set of genotypes, or a subset
(For genomic prediction of new genotypes, see the function nGnE).
Constraints: the genotypic main effect for the first genotype is put to 0.

Penalization: model (refer to equation above) is fitted using glmnet, simultaneously penalizing
\mjeqn{e_j}{ascii}, \mjeqn{g_i}{ascii} and \mjeqn{\beta_{ik}}{ascii}.
If penG = 0 and penE = 0, the main effects \mjeqn{g_i}{ascii} and
\mjeqn{e_j}{ascii} are not penalized. If these parameters are 1, the
the main effects are penalized to the same degree as the sensitivities. Any
nonnegative values are allowed.

Another extension is the option to inlcude common sensitivities that are
the same for each genotype, replacing model (...) by
\mjeqn{Y_{ij} = \mu + e_j + g_i + \sum_{k=1}^s (\beta_{k} + \beta_{ik}) x_{ij} + \epsilon_{ij},}{ascii}
where, for the the kth environmental covariate,
\mjeqn{\beta_{k}}{ascii} is the general and
\mjeqn{\beta_{ik}}{ascii} the genotype-specific sensitivity.

Predictions for the test environments are first constructed using the
estimated genotypic main effects and sensitivities; next, predicted environmental
main effects are added. The latter are obtained by regressing the estimated
environmental main effects for the training environment on the average values
of the indices in these environments, as in Millet et al 2019.
}
