% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GnE_BGLR.R
\name{GnE_BGLR}
\alias{GnE_BGLR}
\title{here comes the title}
\usage{
GnE_BGLR(
  dat,
  Y,
  G,
  E,
  K,
  indices = NULL,
  indicesData = NULL,
  testEnv = NULL,
  corType = c("pearson", "spearman"),
  scaling = c("train", "all", "no"),
  nIter = 50,
  burnIn = 10,
  thin = 1
)
}
\arguments{
\item{dat}{A \code{data.frame} with data from multi-environment trials.
Each row corresponds to a particular genotype in a particular environment.
The data do not need to be balanced, i.e. an environment does not need to
contain all genotypes. \code{dat} should contain the training as well as the
test environments (see testEnv)}

\item{Y}{The trait to be analyzed: either of type character, in which case
it should be one of the column names in \code{dat}, or numeric, in which
case the Yth column of \code{dat} will be analyzed.}

\item{G}{The column in \code{dat} containing the factor genotype: either of
type character, in which case it should be one of the column names in
\code{dat}, or numeric, in which case the Gth column of \code{dat} will be
used.}

\item{E}{The column in \code{dat} containing the factor environment: either
of type character, in which case it should be one of the column names in
\code{dat}, or numeric, in which case the Eth column of \code{dat} will be}

\item{K}{...}

\item{indices}{The columns in \code{dat} containing the environmental
indices (vector of type character). Alternatively, if the indices are always
constant within environments (i.e. not genotype dependent), the
environmental data can also be provided using the argument \code{indicesData}
(see below).}

\item{indicesData}{An optional \code{data.frame} containing environmental
indices (covariates); one value for each environment and index. It should
have the environment names as row names (corresponding to the names
contained in \code{dat$E}); the column names are the indices. If
\code{indices} (see before) is also provided, the latter will be ignored.}

\item{testEnv}{vector (character). Data from these environments are not used
for fitting the model. Accuracy is evaluated for training and test
environments separately. The default is \code{NULL}, i.e. no test
environments, in which case the whole data set is training. It is also
possible that there are test environments, but without any data; in this
case, no accuracy is reported for test environments.}

\item{corType}{type of correlation: Pearson (default) or Spearman rank sum.}

\item{scaling}{determines how the environmental variables are scaled.
"train" : all data (test and training environments) are scaled
using the mean and and standard deviation in the training environments.
"all" : using the mean and standard deviation of all environments.
"no" : No scaling.}

\item{nIter}{...}

\item{burnIn}{...}

\item{thin}{...}
}
\value{
A list with the following elements:
\describe{
\item{yHat}{...}
}
}
\description{
\loadmathjax
here comes the description
}
