% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nGE.R
\name{nGE}
\alias{nGE}
\title{this is the title}
\usage{
nGE(dat, Y, G, E, K)
}
\arguments{
\item{dat}{A \code{data.frame} with data from multi-environment trials.
Each row corresponds to a particular genotype in a particular environment.
The data do not need to be balanced, i.e. an environment does not need to
contain all genotypes. \code{dat} should contain the training as well as the
test environments (see testEnv)}

\item{Y}{The trait to be analyzed: either of type character, in which case
it should be one of the column names in \code{dat}, or numeric, in which
case the Yth column of \code{dat} will be analyzed.}

\item{G}{The column in \code{dat} containing the factor genotype (either
character or numeric).}

\item{E}{The column in \code{dat} containing the factor environment
(either character or numeric).}

\item{K}{kinship matrix, or a list of such matrices. (TO DO: checks)}
}
\value{
A list with the following elements:
\describe{
\item{pred}{...}
\item{Vg}{...},
\item{Ve}{...}
}
}
\description{
\loadmathjax
here comes the description
}
