% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nGnE.R
\name{nGnE}
\alias{nGnE}
\title{here comes the title}
\usage{
nGnE(
  GnEOut,
  K,
  dNew,
  pMethod = c("parametric", "nonparametric"),
  corType = c("pearson", "spearman")
)
}
\arguments{
\item{GnEOut}{...}

\item{K}{...}

\item{dNew}{...}

\item{pMethod}{...}

\item{corType}{type of correlation: Pearson (default) or spearman rank sum}
}
\value{
A list with the following elements:
\describe{
\item{parGenoAll}{...}
\item{predNew}{...},
\item{testAccuracyEnv}{...}
}
}
\description{
\loadmathjax
here comes the description
}
