% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factReg.R
\docType{package}
\name{factReg-package}
\alias{factReg}
\alias{factReg-package}
\title{factReg: Multi-Environment Genomic Prediction with Penalized Factorial Regression}
\description{
Multi-environment genomic prediction for training and test environments using penalized factorial regression. Predictions are made using genotype-specific environmental sensitivities as in Millet et al. (2019) \doi{10.1038/s41588-019-0414-y}.
}
\author{
\strong{Maintainer}: Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl} (\href{https://orcid.org/0000-0002-8673-2514}{ORCID})

Authors:
\itemize{
  \item Kruijer Willem (\href{https://orcid.org/0000-0001-7179-1733}{ORCID})
}

Other contributors:
\itemize{
  \item Vahe Avagyan (\href{https://orcid.org/0000-0001-5838-4850}{ORCID}) [contributor]
}

}
\keyword{internal}
