% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frBLUP.R
\name{frBLUP}
\alias{frBLUP}
\title{Penalized factorial regression on the GBLUPs}
\usage{
frBLUP(
  dat,
  Y,
  G,
  E,
  indices = NULL,
  indicesData = NULL,
  K,
  method = c("factReg", "perGeno"),
  testEnv = NULL,
  type = NULL,
  target = c("nGnE", "GnE"),
  gpWeight = NULL,
  useRes = 2,
  corType = c("pearson", "spearman"),
  partition = NULL,
  alpha = 1,
  lambda = NULL,
  penE = 0,
  scaling = c("train", "all", "no"),
  quadratic = FALSE
)
}
\arguments{
\item{dat}{A \code{data.frame} with data from multi-environment trials.
Each row corresponds to a particular genotype in a particular environment.
The data do not need to be balanced, i.e. an environment does not need to
contain all genotypes. \code{dat} should contain the training as well as the
test environments (see testEnv)}

\item{Y}{The trait to be analyzed: either of type character, in which case
it should be one of the column names in \code{dat}, or numeric, in which
case the Yth column of \code{dat} will be analyzed.}

\item{G}{The column in \code{dat} containing the factor genotype: either of
type character, in which case it should be one of the column names in
\code{dat}, or numeric, in which case the Gth column of \code{dat} will be
used.}

\item{E}{The column in \code{dat} containing the factor environment: either
of type character, in which case it should be one of the column names in
\code{dat}, or numeric, in which case the Eth column of \code{dat} will be}

\item{indices}{The columns in \code{dat} containing the environmental
indices (vector of type character). Alternatively, if the indices are always
constant within environments (i.e. not genotype dependent), the
environmental data can also be provided using the argument \code{indicesData}
(see below).}

\item{indicesData}{An optional \code{data.frame} containing environmental
indices (covariates); one value for each environment and index. It should
have the environment names as row names (corresponding to the names
contained in \code{dat$E}); the column names are the indices. If
\code{indices} (see before) is also provided, the latter will be ignored.}

\item{K}{kinship matrix, or a list of such matrices. (TO DO: checks)}

\item{method}{(TO DO) "factReg" (default) or "perGeno"}

\item{testEnv}{vector (character). Data from these environments are not used
for fitting the model. Accuracy is evaluated for training and test
environments separately. The default is \code{NULL}, i.e. no test
environments, in which case the whole data set is training. It is also
possible that there are test environments, but without any data; in this
case, no accuracy is reported for test environments.}

\item{type}{Character vector of length nrow(dat), which whould indicate the
type of observation, which should be one of "GE", "GnE", "nGE", "nGnE"}

\item{target}{character; should be either "GnE" or "nGnE" (default)}

\item{gpWeight}{vector of length \code{nrow(dat)}; default \code{NULL}}

\item{useRes}{If \code{method} is "perGeno": Indicates whether the
genotype-specific regressions are to be fitted on the residuals of a model
with main effects. 0 means no (i.e. fit the regressions on the original
data). 1: residuals of a model with environmental main effects.
2 (default): genotypic and environmental main effects.}

\item{corType}{type of correlation: Pearson (default) or Spearman rank sum.}

\item{alpha}{Type of penalty, as in glmnet (1 = LASSO, 0 = ridge; in between
= elastic net). Default is 1.}

\item{lambda}{Numeric vector; defines the grid over which the penalty lambda
is optimized in cross validation. Default: NULL (defined by glmnet).
Important special case: lambda = 0 (no penalty).}

\item{penE}{numeric; default 0. If 1, environmental main effects are
penalized. If 0, they are not. Any non negative real number is allowed.}

\item{scaling}{determines how the environmental variables are scaled.
"train" : all data (test and training environments) are scaled
using the mean and and standard deviation in the training environments.
"all" : using the mean and standard deviation of all environments.
"no" : No scaling.}

\item{quadratic}{boolean; default \code{FALSE}. If \code{TRUE}, quadratic
terms (i.e., squared indices) are added to the model.}
}
\value{
A list with the following elements:
\describe{
\item{predTrain}{Vector with predictions for the training set
(to do: Add the factors genotype and environment; make a dataframe)}
\item{predTest}{Vector with predictions for the test set (to do: Add the
factors genotype and environment; make a dataframe). To do: add estimated
environmental main effects, not only predicted environmental main effects}
}
}
\description{
Based on multi-environment field trials, fits the factorial regression model
\deqn{G_{ij} = \mu + e_j + g_i + \sum_{k=1}^s \beta_{ik} x_{ij} + \epsilon_{ij},}
where .... with environmental main effects $e_j$, genotypic main effects \eqn{g_{i}} and
genotype-specific environmental sensitivities \eqn{\beta_{ik}}. See e.g. Millet
et al 2019 and van Eeuwijk / Denis 1996 (?). There are \eqn{s} environmental
indices with values \eqn{x_{ij}}. Optionally, predictions can be made for a set
of test environments, for which environmental indices are available. The new
environments must contain the same set of genotypes, or a subset
(For genomic prediction of new genotypes, the function ...).
Constraints: the first level of the factor genotype is put to 0.
Penalization: model (...) is fitted using glmnet, simultaneously penalizing
\eqn{e_j + g_i} and \eqn{\beta_{ik}}. If penG = FALSE (or penE = FALSE) the
main effects \eqn{g_i} (respectively \eqn{e_j}) are not been penalised.
Predictions for the test environments are first constructed using the
estimated main effects and sensitivities; next, predicted environmental
main effects are added. The latter are obtained by regressing the estimated
environmental main effects for the training environment on the average values
of the indices in these environments, as in Millet et al 2019.
}
