% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GnE_BGLR.R
\name{GnE_BGLR}
\alias{GnE_BGLR}
\title{here comes the title}
\usage{
GnE_BGLR(
  dat,
  Y,
  G,
  E,
  K,
  indices = NULL,
  indicesData = NULL,
  testEnv = NULL,
  corType = c("pearson", "spearman"),
  scaling = c("train", "all", "no"),
  nIter = 50,
  burnIn = 10,
  thin = 1
)
}
\arguments{
\item{K}{...}

\item{nIter}{...}

\item{burnIn}{...}

\item{thin}{...}
}
\value{
A list with the following elements:
\describe{
\item{yHat}{...}
}
}
\description{
\loadmathjax
here comes the description
}
