% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nGE.R
\name{nGE}
\alias{nGE}
\title{this is the title}
\usage{
nGE(dat, Y, G, E, K)
}
\arguments{
\item{K}{kinship matrix, or a list of such matrices. (TO DO: checks)}
}
\description{
\loadmathjax
here comes the description
}
